requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.0
djongo==1.3.6
python-decouple==3.8
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
stripe==7.8.0
celery==5.3.4
redis==5.0.1
